list and map display with json lecture in assets
Add an archive assets, inside there is json. Read json in AllwomenLoader, and display the name in the list. Moreover we achieved to display the women in map by using the coordinate. Now begin to link the WomenActivity with the list and the map (just to send the information from fragments to textview in activity_women, this part hasn't been done yet)